Cleanup InternalDistributedSystemIntegrationTest
